\name{addMarelacMenu}
\alias{remMarelacMenu}
\alias{addMarelacMenu}

\title{Add or remove a special "marelac" menue to the RGui}
\description{
  The function is used to add or remove a user menue "marelac" to the RGui.
}
\usage{
remMarelacMenu()
addMarelacMenu()
}

\details{
  This function works on Windows only. It is called automatically when
  marelacTeching is loaded.
}
\value{
  No value is returned
}

\seealso{\code{\link{sethelp}}}

\keyword{misc}
