\name{sethelp}
\alias{sethelp}

\title{Set the style of the help system on Windows}
\description{
  A new menue is added to switch between the different help systems (Windows, HTML, simple text).
}
\usage{
sethelp(opt = 1)
}

\arguments{
  \item{opt}{Type of the menue system: 1=text, 2=html, 3=windows (chm)}
}
\details{
   This function is available on Windows only. It is normally called by a menue.
}
\value{
  No value is returned.
}
\keyword{utilities}
